Extract controller logic to its own file
